+++ packages/components/src/components/radio-button-wrapper/radio-button-wrapper-styles.ts
@@ -1,7 +1,7 @@
 import type { BreakpointCustomizable, Theme } from '../../types';
-import { addImportantToEachRule, getInvertedThemedColors } from '../../styles';
+import { addImportantToEachRule, getHighContrastColors, getInvertedThemedColors } from '../../styles';
 import { getCheckboxRadioJssStyle } from '../../styles/checkbox-radio-styles';
-import { getCss, mergeDeep } from '../../utils';
+import { getCss, isHighContrastMode, mergeDeep } from '../../utils';
 import type { FormState } from '../../utils/form/form-state';
 import { getInlineSVGBackgroundImage } from '../../utils/svg/getInlineSVGBackgroundImage';
 
@@ -11,7 +11,9 @@ export const getComponentCss = (
   isDisabled: boolean,
   theme: Theme
 ): string => {
-  const checkedIconColor = getInvertedThemedColors(theme).primaryColor.replace(/#/g, '%23');
+  const checkedIconColor = isHighContrastMode
+    ? getHighContrastColors().canvasColor
+    : getInvertedThemedColors(theme).primaryColor.replace(/#/g, '%23');
 
   return getCss(
     mergeDeep(getCheckboxRadioJssStyle(hideLabel, state, isDisabled, false, theme), {