+++ packages/components/src/styles/colors.spec.ts
@@ -1,4 +1,5 @@
-import { getThemedColors } from './colors';
+import { getHighContrastColors, getThemedColors } from './colors';
+import * as a11yUtils from '../utils/a11y/a11y';
 
 describe('getThemedColors()', () => {
   it.each<Parameters<typeof getThemedColors>>([['light'], ['dark']])(
@@ -7,4 +8,16 @@ describe('getThemedColors()', () => {
       expect(getThemedColors(theme)).toMatchSnapshot();
     }
   );
+  it.each<Parameters<typeof getThemedColors>>([['light'], ['dark']])(
+    'should return correct merged high contrast mode scheme colors for theme: %s',
+    (theme) => {
+      Object.defineProperty(a11yUtils, 'isHighContrastMode', { value: true });
+      expect(getThemedColors(theme)).toMatchSnapshot();
+    }
+  );
+});
+describe('getHighContrastColors()', () => {
+  it('should return correct high contrast mode scheme colors', () => {
+    expect(getHighContrastColors()).toMatchSnapshot();
+  });
 });